we happy few

+++ webp-server_test.go
@@ -18,7 +18,7 @@ func TestLoadConfig(t *testing.T) {
 	assert.Equal(t, "./exhaust", c.ExhaustPath)
 	assert.Equal(t, "127.0.0.1", c.Host)
 	assert.Equal(t, "3333", c.Port)
-	assert.Equal(t, float32(80), c.Quality)
+	assert.Equal(t, 80, c.Quality)
 	assert.Equal(t, "./pics", c.ImgPath)
 	assert.Equal(t, []string{"jpg", "png", "jpeg", "bmp"}, c.AllowedTypes)
 }
